% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{$.phasegen.distributions.base.MarginalDemeDistributions}
\alias{$.phasegen.distributions.base.MarginalDemeDistributions}
\alias{[[.phasegen.distributions.base.MarginalDemeDistributions}
\title{Access the marginal distribution of a deme}
\usage{
\method{$}{phasegen.distributions.base.MarginalDemeDistributions}(x, name)

\method{[[}{phasegen.distributions.base.MarginalDemeDistributions}(x, name)
}
\arguments{
\item{x}{The marginal deme distributions, such as \code{coal$sfs$demes}.}

\item{name}{Name of a deme or an attribute.}
}
\value{
The marginal distribution of the deme, or the attribute.
}
\description{
Looks up the distribution of a deme by name, such as \code{coal$sfs$demes$pop_0}, and any other attribute otherwise.
}
\examples{
\dontrun{
pg <- load_phasegen()
coal <- pg$Coalescent(
  n = list(pop_0 = 3L, pop_1 = 2L),
  demography = pg$Demography(
    pop_sizes = list(pop_0 = 1, pop_1 = 1),
    events = c(pg$MigrationRateChange(source = "pop_0", dest = "pop_1", time = 0, rate = 1))
  )
)
coal$sfs$demes$pop_0$mean
coal$sfs$demes[["pop_1"]]$mean
}

}
