% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{load_phasegen}
\alias{load_phasegen}
\title{Load the phasegen library}
\usage{
load_phasegen(install = FALSE)
}
\arguments{
\item{install}{A logical. If TRUE, the function will attempt to run install_phasegen().}
}
\value{
A reference to the 'phasegen' Python library loaded through reticulate.
This reference can be used to access 'phasegen' functionalities.
}
\description{
This function imports the Python package 'phasegen' using the reticulate package
and returns a reference to it, optionally installing it first when \code{install = TRUE}.
}
\examples{
\dontrun{
load_phasegen(install = TRUE)
# now you can use phasegen functionalities as per its API
}

}
\seealso{
\link[reticulate]{import} for importing Python modules in R.
}
