% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{persp.phasegen.distributions.base._JointFunction}
\alias{persp.phasegen.distributions.base._JointFunction}
\title{Draw a joint density or joint CDF as a surface}
\usage{
\method{persp}{phasegen.distributions.base._JointFunction}(x, n_points = NULL, title = NULL, theta = 30, phi = 30, ...)
}
\arguments{
\item{x}{The \code{pdf} or \code{cdf} of a joint distribution.}

\item{n_points}{Number of grid points per axis, \code{NULL} for the default.}

\item{title}{Plot title, \code{NULL} for the default.}

\item{theta, phi}{Viewing angles in degrees.}

\item{...}{Further arguments passed to \code{\link[graphics:persp]{graphics::persp()}}.}
}
\value{
The viewing transformation matrix, invisibly.
}
\description{
Draw a joint density or joint CDF as a surface
}
\examples{
\dontrun{
pg <- load_phasegen()
joint <- pg$Coalescent(n = 8L)$sfs$joint_distribution(1L, 2L)
persp(joint$pdf, title = "Joint density")
persp(joint$cdf, title = "Joint CDF")
}

}
