% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{persp.sfsutils.spectrum.JointSFS}
\alias{persp.sfsutils.spectrum.JointSFS}
\title{Draw a joint SFS as a surface}
\usage{
\method{persp}{sfsutils.spectrum.JointSFS}(
  x,
  pops = c(0, 1),
  title = NULL,
  mask_monomorphic = TRUE,
  theta = 30,
  phi = 30,
  ...
)
}
\arguments{
\item{x}{A \code{JointSFS} object, such as \code{coal$jsfs$mean}.}

\item{pops}{Zero-based indices of the populations on the y-axis and the x-axis.}

\item{title}{Plot title, \code{NULL} for none.}

\item{mask_monomorphic}{Logical, whether to mask the monomorphic corners. Default is \code{TRUE}.}

\item{theta, phi}{Viewing angles in degrees.}

\item{...}{Further arguments passed to \code{\link[graphics:persp]{graphics::persp()}}.}
}
\value{
The viewing transformation matrix, invisibly.
}
\description{
Draws the joint SFS of two populations, marginalised over any further populations.
}
\examples{
\dontrun{
pg <- load_phasegen()
coal <- pg$Coalescent(
  n = list(pop_0 = 3L, pop_1 = 3L),
  demography = pg$Demography(
    pop_sizes = list(pop_0 = 1, pop_1 = 1),
    events = c(pg$MigrationRateChange(source = "pop_0", dest = "pop_1", time = 0, rate = 1))
  )
)
persp(coal$jsfs$mean, title = "Mean joint SFS")
}

}
