% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{persp.sfsutils.spectrum.TwoSFS}
\alias{persp.sfsutils.spectrum.TwoSFS}
\title{Draw a 2-SFS as a surface}
\usage{
\method{persp}{sfsutils.spectrum.TwoSFS}(x, title = NULL, theta = 30, phi = 30, ...)
}
\arguments{
\item{x}{A \code{TwoSFS} object.}

\item{title}{Plot title, \code{NULL} for none.}

\item{theta, phi}{Viewing angles in degrees.}

\item{...}{Further arguments passed to \code{\link[graphics:persp]{graphics::persp()}}.}
}
\value{
The viewing transformation matrix, invisibly.
}
\description{
Draws the segregating entries of a 2-SFS, with entry (i, j) at x = j and y = i.
}
\examples{
\dontrun{
pg <- load_phasegen()
persp(pg$Coalescent(n = 10L)$sfs$cov, title = "Kingman")
}

}
