% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{plot.phasegen.demography.Demography}
\alias{plot.phasegen.demography.Demography}
\title{Plot a demography}
\usage{
\method{plot}{phasegen.demography.Demography}(
  x,
  which = c("all", "pop_sizes", "migration"),
  t = NULL,
  title = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{Demography} object, such as \code{coal$demography}.}

\item{which}{What to draw: \code{"all"} for population sizes and migration rates, \code{"pop_sizes"} or \code{"migration"}.}

\item{t}{Times at which to evaluate the trajectories, \code{NULL} for the default.}

\item{title}{Plot title, \code{NULL} for the default.}

\item{ylab}{Label of the y-axis, \code{NULL} for the default.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot a demography
}
\examples{
\dontrun{
pg <- load_phasegen()
d <- pg$Demography(
  pop_sizes = list(pop_0 = 1, pop_1 = 2),
  events = c(pg$MigrationRateChange(source = "pop_0", dest = "pop_1", time = 0, rate = 1))
)
plot(d)
plot(d, which = "migration")
}

}
