% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{plot.phasegen.distributions.base.QuantileFunction}
\alias{plot.phasegen.distributions.base.QuantileFunction}
\title{Plot a quantile function}
\usage{
\method{plot}{phasegen.distributions.base.QuantileFunction}(
  x,
  n_points = NULL,
  bins = NULL,
  title = NULL,
  label = NULL,
  add = NULL,
  linewidth = 0.5,
  alpha = 1,
  ...
)
}
\arguments{
\item{x}{The \code{quantile} function of a distribution.}

\item{n_points}{Number of grid points, \code{NULL} for the default.}

\item{bins}{SFS bins to draw, \code{NULL} for all polymorphic bins.}

\item{title}{Plot title, \code{NULL} for the default.}

\item{label}{Legend label of the curves, \code{NULL} for none.}

\item{add}{A plot returned by a previous call to draw the curves into, \code{NULL} for a new plot.}

\item{linewidth}{Width of the curves.}

\item{alpha}{Opacity of the curves.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot a quantile function
}
\examples{
\dontrun{
pg <- load_phasegen()
plot(pg$Coalescent(n = 10L)$total_branch_length$quantile, title = "Quantile")
}

}
