% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{plot.sfsutils.spectrum.JointSFS}
\alias{plot.sfsutils.spectrum.JointSFS}
\title{Plot a joint SFS as a heatmap}
\usage{
\method{plot}{sfsutils.spectrum.JointSFS}(
  x,
  pops = c(0, 1),
  title = NULL,
  log_scale = TRUE,
  mask_monomorphic = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{JointSFS} object, such as \code{coal$jsfs$mean}.}

\item{pops}{Zero-based indices of the populations on the y-axis and the x-axis.}

\item{title}{Plot title, \code{NULL} for none.}

\item{log_scale}{Logical, whether to use a logarithmic colour scale. Default is \code{TRUE}.}

\item{mask_monomorphic}{Logical, whether to mask the monomorphic corners. Default is \code{TRUE}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Draws the joint SFS of two populations, marginalised over any further populations.
}
\examples{
\dontrun{
pg <- load_phasegen()
coal <- pg$Coalescent(
  n = list(pop_0 = 3L, pop_1 = 3L),
  demography = pg$Demography(
    pop_sizes = list(pop_0 = 1, pop_1 = 1),
    events = c(pg$MigrationRateChange(source = "pop_0", dest = "pop_1", time = 0, rate = 1))
  )
)
plot(coal$jsfs$mean)
}

}
