% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{plot.sfsutils.spectrum.Spectra}
\alias{plot.sfsutils.spectrum.Spectra}
\title{Plot grouped spectra}
\usage{
\method{plot}{sfsutils.spectrum.Spectra}(
  x,
  title = NULL,
  log_scale = FALSE,
  use_subplots = FALSE,
  show_monomorphic = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{Spectra} object.}

\item{title}{Plot title, \code{NULL} for none.}

\item{log_scale}{Logical, whether to use a logarithmic y-axis. Default is \code{FALSE}.}

\item{use_subplots}{Logical, whether to draw each type in its own panel. Default is \code{FALSE}.}

\item{show_monomorphic}{Logical, whether to include the monomorphic entries 0 and n. Default is \code{FALSE}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot grouped spectra
}
\examples{
\dontrun{
pg <- load_phasegen()
coal <- pg$Coalescent(n = 10L)
plot(pg$Spectra(list(mean = coal$sfs$mean, sd = coal$sfs$std)))
}

}
