% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{plot.sfsutils.spectrum.Spectrum}
\alias{plot.sfsutils.spectrum.Spectrum}
\title{Plot a site-frequency spectrum}
\usage{
\method{plot}{sfsutils.spectrum.Spectrum}(x, title = NULL, log_scale = FALSE, show_monomorphic = FALSE, ...)
}
\arguments{
\item{x}{A spectrum, such as \code{coal$sfs$mean}.}

\item{title}{Plot title, \code{NULL} for none.}

\item{log_scale}{Logical, whether to use a logarithmic y-axis. Default is \code{FALSE}.}

\item{show_monomorphic}{Logical, whether to include the monomorphic entries 0 and n. Default is \code{FALSE}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot a site-frequency spectrum
}
\examples{
\dontrun{
pg <- load_phasegen()
plot(pg$Coalescent(n = 10L)$sfs$mean)
}

}
