% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{plot.sfsutils.spectrum.TwoSFS}
\alias{plot.sfsutils.spectrum.TwoSFS}
\title{Plot a 2-SFS as a heatmap}
\usage{
\method{plot}{sfsutils.spectrum.TwoSFS}(x, title = NULL, max_abs = NULL, ...)
}
\arguments{
\item{x}{A \code{TwoSFS} object.}

\item{title}{Plot title, \code{NULL} for none.}

\item{max_abs}{Largest absolute value of the diverging colour scale, \code{NULL} for the largest absolute entry.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Draws the segregating entries of a 2-SFS, such as \code{coal$sfs$corr} or \code{coal$sfs2$mean}, with entry (i, j) at x = j and
y = i. Covariances and correlations are drawn on a diverging colour scale centred at zero.
}
\examples{
\dontrun{
pg <- load_phasegen()
plot(pg$Coalescent(n = 10L)$sfs$corr)
}

}
