% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{print.phasegen_mutation_config}
\alias{print.phasegen_mutation_config}
\title{Print a mutational configuration}
\usage{
\method{print}{phasegen_mutation_config}(x, ...)
}
\arguments{
\item{x}{A configuration converted from a \code{MutationConfig}.}

\item{...}{Further arguments passed to \code{\link[=print]{print()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a mutational configuration
}
