% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{py_to_r.phasegen.distributions.mutation_configs.MutationConfig}
\alias{py_to_r.phasegen.distributions.mutation_configs.MutationConfig}
\title{Convert a mutational configuration to R}
\usage{
\method{py_to_r}{phasegen.distributions.mutation_configs.MutationConfig}(x)
}
\arguments{
\item{x}{A \code{MutationConfig}, as yielded by \code{coal$sfs$get_mutation_configs()}.}
}
\value{
The counts in bin order, a list of integers as for a plain tuple.
}
\description{
Convert a mutational configuration to R
}
