% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{r_to_py.phasegen_mutation_config}
\alias{r_to_py.phasegen_mutation_config}
\title{Convert a mutational configuration to Python}
\usage{
\method{r_to_py}{phasegen_mutation_config}(x, convert = FALSE)
}
\arguments{
\item{x}{A configuration converted from a \code{MutationConfig}.}

\item{convert}{Ignored. The returned Python object does not convert to R.}
}
\value{
The \code{MutationConfig} of the counts in the layout of \code{x}.
}
\description{
Convert a mutational configuration to Python
}
