% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{phasegen_is_installed}
\alias{phasegen_is_installed}
\title{Check if the \code{phasegen} Python module is installed}
\usage{
phasegen_is_installed()
}
\value{
Logical \code{TRUE} if the \code{phasegen} Python module is installed, otherwise \code{FALSE}.
}
\description{
This function uses the reticulate package to verify if the \code{phasegen} Python
module is currently installed.
}
\examples{
\dontrun{
phasegen_is_installed()  # Returns TRUE or FALSE based on the installation status of phasegen
}

}
