% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasegen.R
\name{plot.phasegen.distributions.base._JointFunction}
\alias{plot.phasegen.distributions.base._JointFunction}
\title{Plot a joint density or joint CDF as a heatmap}
\usage{
\method{plot}{phasegen.distributions.base._JointFunction}(x, n_points = NULL, title = NULL, ...)
}
\arguments{
\item{x}{The \code{pdf} or \code{cdf} of a joint distribution.}

\item{n_points}{Number of grid points per axis, \code{NULL} for the default.}

\item{title}{Plot title, \code{NULL} for the default.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot a joint density or joint CDF as a heatmap
}
\examples{
\dontrun{
pg <- load_phasegen()
joint <- pg$Coalescent(n = 8L)$sfs$joint_distribution(1L, 2L)
plot(joint$pdf, title = "Joint density")
}

}
\seealso{
\code{\link[=persp.phasegen.distributions.base._JointFunction]{persp.phasegen.distributions.base._JointFunction()}} for the surface.
}
